// src/users/models/services/generated/user.generated.ts
export interface RoleResponse {
    id: string;
    name: string;
}

export interface AppResponse {
    appId: string;
    appName: string;
}

export interface AuthResponse {
    application: AppResponse;
    created: string;
}

export interface UserResponse {
    id: string;
    username: string;
    firstName: string;
    lastName: string;
    email: string;
    created: string;
    hasGrantedAccess: boolean;
    grantedAccessOn: string;
    roles: RoleResponse[];
    apps: AppResponse[];
    auths: AuthResponse[];
}

export interface UpdateSelfPasswordRequest {
    currentPassword: string;
    password: string;
    confirmPassword: string;
}

export interface ErrorResponse {
    statusCode: number;
    message: string;
    timestamp: string;
    path: string;
    "x-request-id": string;
}

export interface UsersResponse {
    users: UserResponse[];
    total: number;
    hasNextPage: boolean;
    page: number;
    take: number;
}